// packages/luna-core/src/type.d.ts
export type Schema = z.ZodTypeAny;
export type Schemas = Record<string, Schema>;
export type ZodSchema = ZodObject<{
    [x: string]: Schema;
}, core.$strip>;
export type Nullable<T> = T | null;
export type DataAttributes = {
    [key: `data-${string}`]: string | number | boolean;
};
export type AriaAttributes = {
    [key: `aria-${string}`]: string | number | boolean;
};
export type DataSource = {
    url: string;
    body?: BodyInit | Record<string, unknown>;
    cache?: RequestCache;
    headers?: HeadersInit;
    method?: string;
    namespace?: string;
};
export type Source = {
    [key: string]: DataSource | Array<unknown>;
};
export type Definition = Record<string, unknown>;
export type Orderable = {
    order?: number;
};
export type Hideable = {
    hidden?: boolean;
};
export type Sections = Array<Section>;
export type Value = string | number | readonly string[];
export type Option = {
    description?: string;
    label: string;
    value: string;
};
export type List = {
    advanced?: {
        action?: string;
        length?: Length<number>;
    };
    description?: string;
    fields: Array<Field | Column>;
    label?: string;
    name: string;
    type: 'list' | (string & {});
} & Base;
export type Fields = Array<Field | Column | List>;
export type Base = Orderable & Hideable;
export type CommonProps = {
    disabled?: boolean;
    id?: string;
    name?: string;
    placeholder?: string;
    required?: boolean;
};
export type Section = {
    advanced?: boolean;
    compact?: boolean;
    description?: string;
    fields?: Fields;
    id?: string | number;
    separator?: boolean;
    title?: string;
} & Base;
export type Column = {
    advanced?: {
        cols?: number;
    };
    fields: Array<Field>;
    type: 'column' | (string & {});
} & Base;
export type Operator = 'eq' | 'neq' | 'in' | 'nin' | 'gt' | 'gte' | 'lt' | 'lte';
export type Condition = {
    field?: string;
    operator?: Operator;
    value: string | number | string[];
};
export type CustomValidation = {
    field: string;
    message?: string;
    operator?: Operator;
};
export type FieldState = {
    disabled?: boolean;
    hidden?: boolean;
};
export type ActionEvent<T> = {
    action: T;
};
export type SourceEvent = ActionEvent<'source'> & {
    source: DataSource;
    target: string;
};
export type ValueEvent = ActionEvent<'value'> & {
    value: Record<string, Value | Array<Record<string, unknown>>>;
};
export type StateEvent = ActionEvent<'state'> & {
    state: FieldState;
    target: string | string[];
    when?: string | string[] | Condition;
};
export type ChangeEvent = Array<SourceEvent | ValueEvent | StateEvent>;
export type Validation = {
    custom?: CustomValidation | Array<CustomValidation>;
    email?: string;
    length?: Length<string>;
    required?: string;
};
export type Field = CommonProps & {
    advanced?: {
        aria?: AriaAttributes;
        autocomplete?: string;
        cols?: number;
        data?: DataAttributes;
        entity?: string;
        horizontal?: boolean;
        reverse?: boolean;
    };
    event?: {
        change?: ChangeEvent;
    };
    description?: string;
    fields?: never;
    label?: string;
    name: string;
    readonly?: boolean;
    type: string;
    validation?: Validation;
} & Base;
export type Length<T> = {
    max?: T;
    min?: T;
};
export type Input = Field & {
    advanced?: {
        length?: Length<number>;
    };
};